Reorganize and cleanup
